+++ mods/BrewLAN/units/SSB0401/SSB0401_script.lua
@@ -1,0 +1,66 @@
+--------------------------------------------------------------------------------
+--  Summary:  The Gantry script
+--   Author:  Sean 'Balthazar' Wheeldon
+--------------------------------------------------------------------------------
+local SSeaFactoryUnit = import('/lua/seraphimunits.lua').SSeaFactoryUnit
+--------------------------------------------------------------------------------
+local BrewLANPath = import( '/lua/game.lua' ).BrewLANPath()
+local Buff = import(BrewLANPath .. '/lua/legacy/VersionCheck.lua').Buff
+local GantryUtils = import(BrewLANPath .. '/lua/GantryUtils.lua')
+local BuildModeChange = GantryUtils.BuildModeChange
+local AIStartOrders = GantryUtils.AIStartOrders
+local AIControl = GantryUtils.AIControl
+local AIStartCheats = GantryUtils.AIStartCheats
+local AICheats = GantryUtils.AICheats
+--------------------------------------------------------------------------------
+SSB0401 = Class(SSeaFactoryUnit) {
+    OnCreate = function(self)
+        SSeaFactoryUnit.OnCreate(self)
+        local bp = self:GetBlueprint()
+        self.Rotator1 = CreateRotator(self, 'Pod01', 'y', nil, 5, 0, 0)
+        self.Trash:Add(self.Rotator1)
+
+        self.Rotator2 = CreateRotator(self, 'Pod02', 'y', nil, 8, 0, 0)
+        self.Trash:Add(self.Rotator2)
+
+        self.Rotator3 = CreateRotator(self, 'Pod03', 'y', nil, -3, 0, 0)
+        self.Trash:Add(self.Rotator3)
+
+        self.BuildPointSlider = CreateSlider(self, self:GetBlueprint().Display.BuildAttachBone or 0, -15, 0, 0, -1)
+        self.Trash:Add(self.BuildPointSlider)
+        BuildModeChange(self)
+    end,
+
+    OnKilled = function(self, instigator, type, overkillRatio)
+        self.Rotator1:SetSpeed(0)
+        self.Rotator2:SetSpeed(0)
+        self.Rotator3:SetSpeed(0)
+        SSeaFactoryUnit.OnKilled(self, instigator, type, overkillRatio)
+    end,
+
+    OnStopBeingBuilt = function(self, builder, layer)
+        AIStartCheats(self, Buff)
+        SSeaFactoryUnit.OnStopBeingBuilt(self, builder, layer)
+        AIStartOrders(self)
+    end,
+
+    OnLayerChange = function(self, new, old)
+        SSeaFactoryUnit.OnLayerChange(self, new, old)
+        BuildModeChange(self)
+    end,
+
+    OnStartBuild = function(self, unitBeingBuilt, order)
+        AICheats(self, Buff)
+        SSeaFactoryUnit.OnStartBuild(self, unitBeingBuilt, order)
+        BuildModeChange(self)
+    end,
+
+    OnStopBuild = function(self, unitBeingBuilt)
+        self.UnitControl(self, unitBeingBuilt)
+        SSeaFactoryUnit.OnStopBuild(self, unitBeingBuilt)
+        AIControl(self, unitBeingBuilt)
+        BuildModeChange(self)
+    end,
+}
+
+TypeClass = SSB0401